third_party/blink/web_tests/fast/forms/datetimelocal/datetimelocal-hotkeys.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="../calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="datetime-local" id="datetime-local-0" value="2019-02-14T13:02">
<script>

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-0');
  return openPicker(dateTimeElement)
  .then(() => {
    eventSender.keyDown('PageDown');
    assert_equals(dateTimeElement.value, "2019-03-14T13:02", "Next month hotkey should have incremented month");

    eventSender.keyDown('PageUp');
    assert_equals(dateTimeElement.value, "2019-02-14T13:02", "Previous month hotkey should have decremented month");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Datetimelocal picker: Previous/next month hotkey (PageDown/PageUp)");

</script>
</body>
</html>
